requirements.txt
Flask==3.1.2
matplotlib==3.10.7
PyPDF2==3.0.1
google~=3.0.0
psycopg2==2.9.11
python-dotenv==1.2.1
beautifulsoup4==4.14.2
cryptography~=46.0.3